completed python

+++ linux/python/test.sh
@@ -5,7 +5,10 @@ SHELL_DIR="$( cd "$( dirname "$0" )" && pwd -P )"
 MODULE_NAME=cubrid-python
 SOURCE_DIR=$SHELL_DIR/$MODULE_NAME
 GIT_FILE=$(which git)
+GIT_SOURCE="git@github.com:CUBRID/cubrid-cci.git"
 VERSION_FILE=$SOURCE_DIR/VERSION
+REMOVE_SOURCE=true
+COMMIT_ID=
 
 function show_usage ()
 {
@@ -13,7 +16,9 @@ function show_usage ()
   echo "Note. For ${MODULE_NAME} Driver Release"
   echo ""
   echo " OPTIONS"
-  echo "  -? | -h Show this help message and exit"
+  echo "  -h Show this help message and exit"
+  echo "  -n     Not Remove Source"
+  echo "  -i     Module Commit id"
   echo ""
   echo " Commit-ID"
   echo " Command) git resete --hard [Commit-ID]"
@@ -32,8 +37,8 @@ function test1_python2() {
   mkdir -p python2_result
   mv test_CUBRIDdb.result python2_result/
   mv test_cubrid.result python2_result/
-  #cat python2_result/test_CUBRIDdb.result
-  #cat python2_result/test_cubrid.result
+  cat python2_result/test_CUBRIDdb.result
+  cat python2_result/test_cubrid.result
 }
 
 function test1_python3() {
@@ -43,8 +48,8 @@ function test1_python3() {
   mkdir -p python3_result
   mv test_CUBRIDdb.result python3_result/
   mv test_cubrid.result python3_result/
-  #cat python3_result/test_CUBRIDdb.result
-  #cat python3_result/test_cubrid.result
+  cat python3_result/test_CUBRIDdb.result
+  cat python3_result/test_cubrid.result
 }
 
 function test2() {
@@ -66,27 +71,38 @@ function build () {
 }
 
 function get_source() {
+  echo "Get Source Function"
   if [ "x$GIT_FILE" = "x" ]; then
     echo "[ERROR] Git not found"
     exit 0
   fi
- 
-  if [ "x$SOURCE_DIR" != "x" ]; then
-    echo "Deleted Source"
-    rm -rf $SHELL_DIR/$MODULE_NAME
+  
+  echo "remove source ${REMOVE_SOURCE}"
+  if [ "${REMOVE_SOURCE}" = true ]; then 
+    if [ "x$SOURCE_DIR" != "x" ]; then
+      echo "Deleted Source"
+      rm -rf $SHELL_DIR/$MODULE_NAME
+    fi
+    echo "Source Git Clone"
+    git clone $GIT_SOURCE
+  elif [ ! -d $SOURCE_DIR ]; then
+    echo "Source is not exist"
+    git clone $GIT_SOURCE
   fi
 
-  git clone git@github.com:CUBRID/$MODULE_NAME.git
   cd $SOURCE_DIR
 
-  if [ ! -z $ARG ]; then
-    echo "[CHECK] input commit id : $ARG"
-    git reset --hard $ARG
+  if [ "x$COMMIT_ID" != "x" ]; then
+    echo "[CHECK] input commit id : $COMMIT_ID"
+    git reset --hard $COMMIT_ID
     git submodule init
     git submodule update
   else
-    echo "Get cubrid-cci"
-    git clone git@github.com:CUBRID/cubrid-cci.git cci-src
+    if [ -d $SOURCE_DIR/cci-src ]; then
+      rm -rf $SOURCE_DIR/cci-src
+      echo "Get cubrid-cci"
+      git clone git@github.com:CUBRID/cubrid-cci.git cci-src
+    fi
   fi
 
 
@@ -99,10 +115,14 @@ function get_source() {
 }
 
 echo "SOURCE DIR $SOURCE_DIR"
-
-while getopts "h" opt; do
+while getopts "i:hr" opt; do
   case $opt in
-    h|\?|* ) show_usage; exit 1;;
+    n ) REMOVE_SOURCE=false  
+        echo "set remove source";;
+    i ) echo "commit id = $OPTARG"
+        COMMIT_ID="$OPTARG";;
+    h ) show_usage; exit 1;;
+    * ) echo "$opt is not the option";;
   esac
 done
 
@@ -113,6 +133,6 @@ fi
 
 get_source
 build
-#test1_python2
-#test1_python3
+test1_python2
+test1_python3
 test2
